Refactoring button with icons + review fixes
Review fixes from the pull request

The border in button group is changed to px instead of em because of
issues with rounding errors on some browsers.

CSS properties are reordered and respaced according to the front-end
guidelines

Buttons with icons are refactored again to keep them atoms. More work
will have to be done to keep the icons.