Renamed and fixed icon-with-button
There was confusion with .a-btn_left-icon and .a-btn_icon__left so I
talked with @marteki and changed the positioning classes to
.a-btn_icon-position-left to make it clearer what they do and better
contrast against icons that happen to be named "left" or "right"

I also fixed the css for right positioned icons and made it so the icons
only appear when paired with a icon positioning class to avoid
mishaps.

Fixing breaks in .a-btn-grouped__combined

The lines seperating the buttons in the combined button group were
inconsistent. This fixes it for a group with up to four choices.

code edit: typo in code

Grouped Button Border Fix