Pulling big-checkboxes from Cfgov-refresh
This pull request (https://github.com/cfpb/cfgov-refresh/pull/2292)
added big checkboxes to cfgov-refresh. A lot of decisions on the
structure were made in that PR.

This pulls all that work into capital framework. Once I pulled it in, I
realized some work needs to be done with the colors and variables, so
I'll work on that in the next commit.