Big atomic update to cf-icons
Default usage of icons required an empty elem because the font family was set on the elem level. This change restricts the font-family to the pseudo-elem allowing icons to be used with standard text. This change will eliminate the duplicated icon classes in cf-buttons and cf-typography.

__NOTE:__

The default :before state still exists to avoid having to update every icon usage at this time but the preferred usage will still be to specify a before or after modifier for any new work going forward.