Fix a bug when animating icons within a button
Currently, if you add the `cf-icon__spin` class to an icon within a button, it spins everything, the padding, border, and margin due to those properties being attached to the icon pseudo element. In order to spin the icon without spinning the other properties, we have to separate the two. This change won't affect icon usage anywhere else, only this specific case (meaning we would only need the empty spans here).