Add --dry-run option to simulate npm release
Travis' release script is admittedly a black box. It's hard to know
exactly which components will be published to npm because running the
release script locally results in components being tagged and permanently
uploaded to npm. This new dry run option allows the script to be run
without doing any permanent damage.