Update the browser list config and support documentation
As part of the FEWD Browser Support Working Group we have been
re-evaluating what browsers we're supporting. We had agreed to use a
browserlistrc file to track our needs, but as we looked closer we
realized our build process is broken up by browser and would necessitate
a complex browserlistrc setup.

- Converted browser config to a new separate config file.
- Reduced configs to only those necessary (many browsers are included in
  "last2").
- Added human readable browser support notes to contributing doc.