Test the release scripts
Set up a gulp test task to test the release scripts and wrote unit tests
for most of the release files. There are a few gaps that we should
return to in the future. There are also optimizations to make in the
release scripts in the future, specifically not relying on process.