Prevent release scripts from being triggered by CFPBot
The echo'ing of environment variables was moved to the top of the
release script so that they'll be printed even if the script doesn't
abort.

The script now checks if the last git commit's username is CFPBot. If
so, it aborts. The git username is set in the same file (L32) so even
if someone changes the bot's GH username or an entirely different bot
is used, the conditional will still operate as expected.

https://mattermost/cfpb/pl/79thwrarg3bhbr6yxp9ygcn8zh