Fix the versions of cf-buttons dependency
When v5.0.0 of cf-buttons were released, the other
components that depended on it needed to have their manifests updated
to match the new major version. Right now the bot doesn't know anything
about this and doesn't take it into account with major releases. For the
time being, any time we publish a major release, we'll need to follow it
up with a patch to update any sibling packages. Long term we should look
into upgrading the bot to figure it out and automate it.