Fix the suggested markup of Notifications
Our in-use Notifications are currently setting the explanation content
as `h4` requiring a reset in the styles. Additionally the content is
wrapped in a `div` rather than the more appropriate `p` requiring
additional styles for any links within. This change fixes both of these
issues as well as includes the explanation markup in the doc examples
for reference.