Bump cf-core due to strange phantom publish
At 2018-08-20T16:51:53.111Z version 4.8.0 of cf-core was published
shortly after https://github.com/cfpb/capital-framework/pull/855 was
opened. We're not sure how this happened (Travis logs are all clean).
To play it safe we're deleting cf-core@4.8.0 and I'm bumping its
package.json to 4.8.0 so that the release PR will put it at 4.9.0.

https://[chat]/cfpb/pl/p4pjk7x9qbyx8fsdcoedkxu4ke